+++ CHANGELOG.md
@@ -7,6 +7,11 @@ A major feature overhaul.
 ### Added
 
 - **TTT2** Bots will show their avatars in the top-right when their corpse is revealed. (#41)
+- Added concommand `ttt_bot_nav_gen` to utilzie Navmesh Optimizer for quick navmesh generation.
+
+### Removed
+
+- Removed concommand `ttt_bot_nav_generate`, as it was outdated.
 
 ## v1.2.2